#  [PT Run] Improve the UWP Program Indexing speed