--- a/src/modules/launcher/Plugins/Microsoft.Plugin.Program/Programs/AppxPackageHelper.cs
+++ b/src/modules/launcher/Plugins/Microsoft.Plugin.Program/Programs/AppxPackageHelper.cs
@@ -11,12 +11,13 @@ namespace Microsoft.Plugin.Program.Programs
 {
     public static class AppxPackageHelper
     {
+        private static readonly IAppxFactory AppxFactory = (IAppxFactory)new AppxFactory();
+
         // This function returns a list of attributes of applications
         public static List<IAppxManifestApplication> GetAppsFromManifest(IStream stream)
         {
             List<IAppxManifestApplication> apps = new List<IAppxManifestApplication>();
-            var appxFactory = new AppxFactory();
-            var reader = ((IAppxFactory)appxFactory).CreateManifestReader(stream);
+            var reader = AppxFactory.CreateManifestReader(stream);
             var manifestApps = reader.GetApplications();
 
             while (manifestApps.GetHasCurrent())