--- a/src/modules/launcher/Plugins/Microsoft.Plugin.Program/Programs/PackageWrapper.cs
+++ b/src/modules/launcher/Plugins/Microsoft.Plugin.Program/Programs/PackageWrapper.cs
@@ -52,9 +52,7 @@ namespace Microsoft.Plugin.Program.Programs
             string path;
             try
             {
-                path = IsPackageDotInstallationPathAvailable.Value
-                    ? package.InstalledPath
-                    : package.InstalledLocation.Path;
+                path = IsPackageDotInstallationPathAvailable.Value ? GetInstalledPath(package) : package.InstalledLocation.Path;
             }
             catch (Exception e) when (e is ArgumentException || e is FileNotFoundException || e is DirectoryNotFoundException)
             {
@@ -76,5 +74,9 @@ namespace Microsoft.Plugin.Program.Programs
                     package.IsDevelopmentMode,
                     path);
         }
+
+        // This is a seperate method so the reference to .InstalledPath won't be loaded in API versions which do not support this API (e.g. older then Build 19041)
+        private static string GetInstalledPath(Package package)
+            => package.InstalledPath;
     }
 }