--- a/src/modules/launcher/Plugins/Microsoft.Plugin.Program/Programs/UWP.cs
+++ b/src/modules/launcher/Plugins/Microsoft.Plugin.Program/Programs/UWP.cs
@@ -1,4 +1,4 @@
-// Copyright (c) Microsoft Corporation
+﻿// Copyright (c) Microsoft Corporation
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -20,6 +20,13 @@ namespace Microsoft.Plugin.Program.Programs
     public partial class UWP
     {
         private static readonly IPath Path = new FileSystem().Path;
+
+        private static readonly Dictionary<string, PackageVersion> _versionFromNamespace = new Dictionary<string, PackageVersion>
+        {
+            { "http://schemas.microsoft.com/appx/manifest/foundation/windows10", PackageVersion.Windows10 },
+            { "http://schemas.microsoft.com/appx/2013/manifest", PackageVersion.Windows81 },
+            { "http://schemas.microsoft.com/appx/2010/manifest", PackageVersion.Windows8 },
+        };
 
         public string Name { get; }
 
@@ -61,16 +68,7 @@ namespace Microsoft.Plugin.Program.Programs
 
             if (hResult == Hresult.Ok)
             {
-                var apps = new List<UWPApplication>();
-
-                List<IAppxManifestApplication> appsViaManifests = AppxPackageHelper.GetAppsFromManifest(stream);
-                foreach (var appInManifest in appsViaManifests)
-                {
-                    var app = new UWPApplication(appInManifest, this);
-                    apps.Add(app);
-                }
-
-                Apps = apps.Where(a =>
+                Apps = AppxPackageHelper.GetAppsFromManifest(stream).Select(appInManifest => new UWPApplication(appInManifest, this)).Where(a =>
                 {
                     var valid =
                     !string.IsNullOrEmpty(a.UserModelId) &&
@@ -90,7 +88,7 @@ namespace Microsoft.Plugin.Program.Programs
                 var e = Marshal.GetExceptionForHR((int)hResult);
                 ProgramLogger.Exception("Error caused while trying to get the details of the UWP program", e, GetType(), path);
 
-                Apps = new List<UWPApplication>().ToArray();
+                Apps = Array.Empty<UWPApplication>();
             }
         }
 
@@ -118,20 +116,10 @@ namespace Microsoft.Plugin.Program.Programs
 
         private void InitPackageVersion(string[] namespaces)
         {
-            var versionFromNamespace = new Dictionary<string, PackageVersion>
+            foreach (var n in _versionFromNamespace.Keys.Where(namespaces.Contains))
             {
-                { "http://schemas.microsoft.com/appx/manifest/foundation/windows10", PackageVersion.Windows10 },
-                { "http://schemas.microsoft.com/appx/2013/manifest", PackageVersion.Windows81 },
-                { "http://schemas.microsoft.com/appx/2010/manifest", PackageVersion.Windows8 },
-            };
-
-            foreach (var n in versionFromNamespace.Keys)
-            {
-                if (namespaces.Contains(n))
-                {
-                    Version = versionFromNamespace[n];
-                    return;
-                }
+                Version = _versionFromNamespace[n];
+                return;
             }
 
             ProgramLogger.Exception($"|Trying to get the package version of the UWP program, but a unknown UWP appmanifest version {FullName} from location {Location} is returned.", new FormatException(), GetType(), Location);
@@ -162,11 +150,10 @@ namespace Microsoft.Plugin.Program.Programs
                     }
 
                     return u.Apps;
-                }).ToArray();
+                }).ToList();
 
                 var updatedListWithoutDisabledApps = applications
-                                                        .Where(t1 => !Main.Settings.DisabledProgramSources
-                                                                        .Any(x => x.UniqueIdentifier == t1.UniqueIdentifier))
+                                                        .Where(t1 => Main.Settings.DisabledProgramSources.All(x => x.UniqueIdentifier != t1.UniqueIdentifier))
                                                         .Select(x => x);
 
                 return updatedListWithoutDisabledApps.ToArray();
@@ -180,26 +167,20 @@ namespace Microsoft.Plugin.Program.Programs
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Intentionally keeping the process alive.")]
         private static IEnumerable<IPackage> CurrentUserPackages()
         {
-            var ps = PackageManagerWrapper.FindPackagesForCurrentUser();
-            ps = ps.Where(p =>
+            return PackageManagerWrapper.FindPackagesForCurrentUser().Where(p =>
             {
-                bool valid;
                 try
                 {
                     var f = p.IsFramework;
                     var path = p.InstalledLocation;
-                    valid = !f && !string.IsNullOrEmpty(path);
+                    return !f && !string.IsNullOrEmpty(path);
                 }
                 catch (Exception e)
                 {
                     ProgramLogger.Exception("An unexpected error occurred and unable to verify if package is valid", e, MethodBase.GetCurrentMethod().DeclaringType, "id");
                     return false;
                 }
-
-                return valid;
             });
-
-            return ps;
         }
 
         public override string ToString()