## Summary of the Pull Request

**What is this about:**
after PR #11364 i decided to take a check on the UWP side as well (less resources used there means it's faster and more resources are available for other plugins as well :)

**What is include in the PR:** 

- Changed  package.InstalledLocation.Path to package.InstalledPath (seems to be 40x faster)
- Changed GetAppsFromManifest to an Enumerator (So we reduce intermediate list allocations)
- Use Enumerable.Empty<>/ Array.Empty<> for null/empty results instead of new List<>/new Array<>
- Moved Lookup lists outside the method
- Change  File.ReadAllBytes(path) to File.OpenRead()
   - Directly read to the output instead of using a temporary bytearray
- Small code readability changes
- Fixed a bug where we were using a duplicate name inside a method / instance variable

**How does someone test / validate:** 

## Quality Checklist

- [x] **Linked issue:** #7208
- [ ] **Communication:** I've discussed this with core contributors in the issue. 
- [x] **Tests:** Added/updated and all pass
- [ ] **Installer:** Added/updated and all pass
- [ ] **Localization:** All end user facing strings can be localized
- [ ] **Docs:** Added/ updated
- [ ] **Binaries:** Any new files are added to WXS / YML
   - [x] No new binaries
   - [ ] [YML for signing](https://github.com/microsoft/PowerToys/blob/master/.pipelines/pipeline.user.windows.yml#L68) for new binaries
   - [ ] [WXS for installer](https://github.com/microsoft/PowerToys/blob/master/installer/PowerToysSetup/Product.wxs) for new binaries

## Contributor License Agreement (CLA)
A CLA must be signed. If not, go over [here](https://cla.opensource.microsoft.com/microsoft/PowerToys) and sign the CLA.

#Perf results
I've build the solution as 'DEBUG", ran it 4 times and skipped the first sample.

## Branch
Microsoft.Plugin.Program.Main - Package index cost <1947ms>
Microsoft.Plugin.Program.Main - Package index cost <1956ms>
Microsoft.Plugin.Program.Main - Package index cost <2127ms>

## Master
Microsoft.Plugin.Program.Main - Package index cost <2592ms>
Microsoft.Plugin.Program.Main - Package index cost <2767ms>
Microsoft.Plugin.Program.Main - Package index cost <2511ms>