@mykhailopylyp 2021-06-10
Measured improvement.
On the branch
```
[2021-06-10 12:20:37.1165] [INFO] [C:\JaneaSystems\PowerToysOther\PowerToys.royvou\src\modules\launcher\Wox.Infrastructure\Stopwatch.cs::50]
Microsoft.Plugin.Program.Main - Package index cost <688ms>

[2021-06-10 12:21:21.9628] [INFO] [C:\JaneaSystems\PowerToysOther\PowerToys.royvou\src\modules\launcher\Wox.Infrastructure\Stopwatch.cs::50]
Microsoft.Plugin.Program.Main - Package index cost <659ms>

[2021-06-10 12:21:47.1876] [INFO] [C:\JaneaSystems\PowerToysOther\PowerToys.royvou\src\modules\launcher\Wox.Infrastructure\Stopwatch.cs::50]
Microsoft.Plugin.Program.Main - Package index cost <630ms>
```

On the master branch
```
[2021-06-10 12:29:15.5332] [INFO] [C:\JaneaSystems\PowerToysOther\PowerToys.royvou\src\modules\launcher\Wox.Infrastructure\Stopwatch.cs::50]
Microsoft.Plugin.Program.Main - Package index cost <1490ms>

[2021-06-10 12:30:11.2012] [INFO] [C:\JaneaSystems\PowerToysOther\PowerToys.royvou\src\modules\launcher\Wox.Infrastructure\Stopwatch.cs::50]
Microsoft.Plugin.Program.Main - Package index cost <1493ms>

[2021-06-10 12:30:25.3017] [INFO] [C:\JaneaSystems\PowerToysOther\PowerToys.royvou\src\modules\launcher\Wox.Infrastructure\Stopwatch.cs::50]
Microsoft.Plugin.Program.Main - Package index cost <1581ms>
```